+++ frontend/src/assets/scoreExplanations.ts
